+++ geometry/generate_random_points.h
@@ -16,7 +16,7 @@
 /// This function generates the given number of points within the 
 /// given bounding box, from the given probability density function. The 
 /// given random number generator is used.
-void generate_random_points(rng_t* rng, sp_func_t* density, bbox_t* bounding_box, int num_points, point_t* points);
+void generate_random_points(rng_t* rng, sp_func_t* density, bbox_t* bounding_box, size_t num_points, point_t* points);
 
 ///@}
